requirements.txt
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.3
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.7.0
